Add Motion loading and move it to core
