+++ src/spikeinterface/benchmark/benchmark_motion_estimation.py
@@ -8,15 +8,15 @@ import time
 import numpy as np
 
 from spikeinterface.core import get_noise_levels
-from .benchmark_base import Benchmark, BenchmarkStudy
-from .benchmark_plot_tools import _simpleaxis
+from spikeinterface.core.motion import Motion
 from spikeinterface.sortingcomponents.motion import estimate_motion
 from spikeinterface.sortingcomponents.peak_detection import detect_peaks
 from spikeinterface.sortingcomponents.peak_selection import select_peaks
 from spikeinterface.sortingcomponents.peak_localization import localize_peaks
 from spikeinterface.widgets import plot_probe_map
 
-from spikeinterface.sortingcomponents.motion import Motion
+from .benchmark_base import Benchmark, BenchmarkStudy
+from .benchmark_plot_tools import _simpleaxis
 
 # import MEArec as mr
