+++ src/spikeinterface/sortingcomponents/motion/__init__.py
@@ -1,4 +1,3 @@
-from .motion_utils import Motion
 from .motion_estimation import estimate_motion
 from .motion_interpolation import (
     correct_motion_on_peaks,
