+++ src/spikeinterface/widgets/motion.py
@@ -5,7 +5,7 @@ import numpy as np
 from .base import BaseWidget, to_attr
 
 from spikeinterface.core import BaseRecording, SortingAnalyzer
-from spikeinterface.sortingcomponents.motion import Motion
+from spikeinterface.core.motion import Motion
 
 
 class MotionWidget(BaseWidget):
